refactor(http): extract the connection pool into its own translation unit
aether_http_server.c mixes HTTP/1.1 parsing, routing, TLS, WebSocket,
SSE, static-file serving, h2 upgrade and thread-budget management in one
4,759-line file. Thread-budget management is the cleanest seam: the pool
needs nothing from the server internals beyond HttpServer and the
already-public http_server_drain_connection, so it moves out without an
internal header. handle_client_connection was only a wrapper over that
public helper, so the pool now calls it directly.

Pure move plus that one call swap; no behaviour change. Verified by a
clean rebuild, a MinGW cross-compile of both files, and the HTTP/2,
reverse-proxy and keep-alive integration suites.

This is one seam, not the whole split: the file is still 4,618 lines and
still carries the other responsibilities listed above.
